fix test last exception error in tests.
The problem is that now an early exception might be raised by bin/test
when trying to create a test directory which might already exist.

This exception is captured but is Python's last exception, which somehow
gets into a traceback in our tests.

This "fix" just makes the tests less specific about the about, so that we
can ignore this traceback.